Add context category (Part 3)
* Moved extracted views to their own files
* Disabled add button when there is no text in the text field

Fix #61
